done with palindrome

+++ lab02/src/adventure/SpeciesListStage.java
@@ -97,7 +97,7 @@ public class SpeciesListStage implements AdventureStage {
      * double in [0, 1]
      */
     public static double arraySimilarity(List<String> listOne, List<String> listTwo) {
-        List<String> copy = new ArrayList<>(listOne);
+        List<String> copy =  new ArrayList<>(listOne);
         int similarObjects = 0;
         for (String o : listTwo) {
             if (copy.contains(o)) {
@@ -105,6 +105,6 @@ public class SpeciesListStage implements AdventureStage {
                 copy.remove(o);
             }
         }
-        return similarObjects / listOne.size();
+        return similarObjects /(float)listOne.size();
     }
 }
